:joy_cat: Update the license to Eclipse Public License - v 2.0
This commit changes the license of the software to the Eclipse Public License - v 2.0. The new license grants certain rights and responsibilities to contributors and recipients of the program. It defines terms such as "Contribution," "Contributor," "Licensed Patents," and "Derivative Works." The grant of rights includes copyright and patent licenses, allowing recipients to reproduce, prepare derivative works, publicly display, perform, distribute, and sublicense the contributions. The license also sets requirements for distributing the program in source code form and prohibits the removal or alteration of certain notices. Additionally, the license addresses commercial distribution and indemnification obligations for contributors. The program is provided "as is" without any warranties or conditions. Recipients assume all risks associated with using and distributing the program. The license disclaims liability for any damages, including lost profits, arising from the use or distribution of the program.
Signed-off-by: Ionut I <git@h-all.co>